layout/base/tests/test_mozPaintCount.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tests for mozPaintCount</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="plugin-utils.js"></script>
  <script type="application/javascript">
    setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);
  </script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="doBackgroundFlicker()">
<p id="display">
<embed type="application/x-test" width="100" height="100" id="p"
       drawmode="solid" color="FF00FF00"></embed>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var startPaintCount = window.mozPaintCount;
ok(true, "Got to initial paint count: " + startPaintCount);
var color = 0;

function doPluginFlicker() {
  ok(true, "Plugin color iteration " + color + ", paint count: " + window.mozPaintCount);
  if (window.mozPaintCount - startPaintCount > 20) {
    ok(true, "Got enough paints from plugin color changes");
    SimpleTest.finish();
    return;
  }

  color = (color + 1) % 256;
  var str = color.toString(16);
  if (str.length < 2) {
    str = "0" + str;
  }
  str = "FF" + str + str + str;
  document.getElementById("p").setColor(str);
  setTimeout(doPluginFlicker, 0);
}

function doBackgroundFlicker() {
  ok(true, "Background color iteration " + color + ", paint count: " + window.mozPaintCount);
  if (window.mozPaintCount - startPaintCount > 20) {
    ok(true, "Got enough paints from background color changes");
    startPaintCount = window.mozPaintCount;
    doPluginFlicker();
    return;
  }

  color = (color + 1) % 256;
  document.body.style.backgroundColor = "rgb(" + color + "," + color + "," + color + ")";
  setTimeout(doBackgroundFlicker, 0);
}

</script>
</pre>

<div style="height:4000px"></div>
<a id="first"  href="http://www.mozilla.org/">first<br>link</a>
<a id="second" href="http://www.mozilla.org/">second link</a>
<a id="third"  href="http://www.mozilla.org/">third<br>link</a>
<div style="height:4000px"></div>

</body>
</html>
